refactor: remove unused DaoDashboardTabs component and refactor dashboard to include new pending DAO member applications card
